I refactored the BrickViewHolder class to be under the new viewholder subpackage
