+++ BrickKit/app/src/main/java/com/wayfair/brickkitdemo/bricks/UnusedBrick.java
@@ -8,7 +8,7 @@ import android.view.View;
 import com.wayfair.brickkit.brick.BaseBrick;
 import com.wayfair.brickkit.padding.BrickPadding;
 import com.wayfair.brickkit.size.BrickSize;
-import com.wayfair.brickkit.BrickViewHolder;
+import com.wayfair.brickkit.viewholder.BrickViewHolder;
 import com.wayfair.brickkitdemo.R;
 
 /**
