+++ BrickKit/bricks/src/androidTest/java/com/wayfair/brickkit/behavior/StickyFooterBehaviorTest.java
@@ -13,7 +13,7 @@ import android.widget.TextView;
 
 import com.wayfair.brickkit.BrickDataManager;
 import com.wayfair.brickkit.BrickRecyclerAdapter;
-import com.wayfair.brickkit.BrickViewHolder;
+import com.wayfair.brickkit.viewholder.BrickViewHolder;
 import com.wayfair.brickkit.R;
 import com.wayfair.brickkit.StickyScrollMode;
 import com.wayfair.brickkit.brick.BaseBrick;
