+++ BrickKit/bricks/src/androidTest/java/com/wayfair/brickkit/behavior/StickyHeaderBehaviorTest.java
@@ -16,7 +16,7 @@ import android.widget.TextView;
 
 import com.wayfair.brickkit.BrickDataManager;
 import com.wayfair.brickkit.BrickRecyclerAdapter;
-import com.wayfair.brickkit.BrickViewHolder;
+import com.wayfair.brickkit.viewholder.BrickViewHolder;
 import com.wayfair.brickkit.R;
 import com.wayfair.brickkit.StickyScrollMode;
 import com.wayfair.brickkit.brick.BaseBrick;
