+++ BrickKit/bricks/src/androidTest/java/com/wayfair/brickkit/util/BrickTestHelper.java
@@ -6,7 +6,7 @@ package com.wayfair.brickkit.util;
 import android.content.Context;
 import android.view.View;
 
-import com.wayfair.brickkit.BrickViewHolder;
+import com.wayfair.brickkit.viewholder.BrickViewHolder;
 import com.wayfair.brickkit.brick.BaseBrick;
 import com.wayfair.brickkit.padding.BrickPadding;
 import com.wayfair.brickkit.padding.SimpleBrickPadding;
