+++ BrickKit/bricks/src/androidTest/java/com/wayfair/brickkit/viewholder/factory/BrickViewHolderFactoryTest.java
@@ -8,7 +8,7 @@ import android.widget.FrameLayout;
 import androidx.test.platform.app.InstrumentationRegistry;
 import androidx.test.runner.AndroidJUnit4;
 
-import com.wayfair.brickkit.BrickViewHolder;
+import com.wayfair.brickkit.viewholder.BrickViewHolder;
 import com.wayfair.brickkit.test.R;
 import com.wayfair.brickkit.brick.BaseBrick;
 import com.wayfair.brickkit.brick.BrickProvider;
