+++ BrickKit/bricks/src/main/java/com/wayfair/brickkit/viewholder/BrickViewHolder.java
@@ -1,18 +1,19 @@
 /**
  * Copyright © 2017 Wayfair. All rights reserved.
  */
-package com.wayfair.brickkit;
+package com.wayfair.brickkit.viewholder;
 
 import android.view.View;
 
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.wayfair.brickkit.BrickRecyclerAdapter;
 
 /**
  * Base ViewHolder for bricks. It extends {@link androidx.recyclerview.widget.RecyclerView.ViewHolder} with
  * one additional method that is called when the view holder is detached from the {@link RecyclerView} so
  * that views can be released.
  */
-//TODO: migrate this class to the view holder package
 public class BrickViewHolder extends RecyclerView.ViewHolder {
     /**
      * Constructor.
@@ -27,13 +28,13 @@ public class BrickViewHolder extends RecyclerView.ViewHolder {
      * Called when {@link BrickRecyclerAdapter#onViewAttachedToWindow(BrickViewHolder)}
      * is invoked.
      */
-    protected void onViewAttachedToWindow() {
+    public void onViewAttachedToWindow() {
     }
 
     /**
      * Method called when the view is detached from the {@link RecyclerView}. All views that have resources
      * should release them here (e.g. ImageView).
      */
-    protected void releaseViewsOnDetach() {
+    public void releaseViewsOnDetach() {
     }
 }
