+++ BrickKit/bricks/src/main/java/com/wayfair/brickkit/viewholder/factory/BrickViewHolderFactoryData.java
@@ -4,7 +4,7 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 
-import com.wayfair.brickkit.BrickViewHolder;
+import com.wayfair.brickkit.viewholder.BrickViewHolder;
 import com.wayfair.brickkit.brick.BrickProvider;
 
 /**
